gis_fetcher/requirements.txt
rasterio
numpy
pandas
requests
geopandas
pyproj
python-dotenv
